Add ForfeitTx domain & tidy up scaffolding
